refactor: make tailwind nav the default and update tests
