+++ packages/app-react/src/components/Nav/Nav.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import { render, cleanup } from '@testing-library/react';
+import { render, cleanup, getByText } from '@testing-library/react';
 import { AuthStateContext } from '../../context/authContext';
 import Nav from '.';
 
@@ -24,8 +24,8 @@ describe('Nav component', () => {
     };
 
     it('renders a nav element', () => {
-        const { container } = renderWithContext({ contextUser: false, isFetching: false });
-        expect(container.firstChild.nodeName).toBe('NAV');
+        const { getByTestId } = renderWithContext({ contextUser: false, isFetching: false });
+        expect(getByTestId('nav')).toBeTruthy();
     });
 
     it('has h3 element with logo text', () => {
@@ -34,17 +34,17 @@ describe('Nav component', () => {
     });
 
     it('has Signup link if user is not logged in', () => {
-        const { getByText } = renderWithContext({ contextUser: false, isFetching: false });
-        expect(getByText('Signup').nodeName).toBe('A');
+        const { getByTestId } = renderWithContext({ contextUser: false, isFetching: false });
+        expect(getByTestId('desktop-signup')).toBeTruthy();
     });
 
     it('has Login link if user is not logged in', () => {
-        const { getByText } = renderWithContext({ contextUser: false, isFetching: false });
-        expect(getByText('Login').nodeName).toBe('A');
+        const { getByTestId } = renderWithContext({ contextUser: false, isFetching: false });
+        expect(getByTestId('desktop-login')).toBeTruthy();
     });
 
     it('has Logout button if user is logged in', () => {
-        const { getByText } = renderWithContext({ contextUser: authtUser, isFetching: false });
-        expect(getByText('Logout').nodeName).toBe('BUTTON');
+        const { getByTestId } = renderWithContext({ contextUser: authtUser, isFetching: false });
+        expect(getByTestId('desktop-logout')).toBeTruthy();
     });
 });
